unauthorized exception now comes from service class, not database. most tests pass for phase 3

+++ server/src/main/java/service/UserService.java
@@ -1,6 +1,7 @@
 package service;
 
 import dataaccess.DataAccess;
+import dataaccess.DataAccessException;
 import datamodel.AuthData;
 import datamodel.UserData;
 
@@ -38,7 +39,11 @@ public class UserService {
     }
 
     public void logout(String authToken) {
+        if (dataAccess.getAuth(authToken) == null) {
+            throw new UnauthorizedException("Not a recognized auth token");
+        }
         dataAccess.deleteAuth(authToken);
+
     }
 
     public void clear() {
